Update patch set 2
Patch Set 2:

Round 2.   All of the framework files were removed, and the docs directory for read the docs stuff now has placeholders and the python goo that the scraper needs.  Docker file is trimmed  a bit, and the project builds with CMake as that is easiest for the CI stuff.

Patch-set: 2
Reviewer: Gerrit User 1000049 <1000049@84e4e190-e623-4c37-9eda-58b1937c1b53>