Update TS to enforce A1 policy on throughput predictions
TS xApp does not enforce the A1 Policy "threshold" to send CONTROL
messages. This feature was disabled in previous changes to allow
interoperability with other xApps in the TS use case.

This change re-enables the "threshold" policy in TS xApp.
If one of the cells has a higher throughput prediction above
a given "threshold" percentage than the current serving cell,
then TS sends a CONTROL request to hanfoff the UE to the cell
with the higher throughput.

Issue-ID: RICAPP-188

Signed-off-by: Alexandre Huff <alexandrehuff@utfpr.edu.br>
Change-Id: I1bdaa80407aee6873107ea4c70e1781df6bbb7af